feat: post and get refactor
